tts-server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
chatterbox-tts>=0.1.0
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
scipy>=1.11.0
numpy>=1.24.0
